[cleanup] clean indents/semicolons, rename api.js to apiRouter.js
