+++ instatrip/config.js
@@ -1,0 +1,3 @@
+var clientID = 'ea354b6fc30e4b1180efe33ed233f7fe';
+var secret = '01397a2619064c898dc1ad0f72bc619e';
+var gmaps = 'AIzaSyA7TbsUFsundVA54vNphHBB3Vn9vmxQYBs';
