+++ npm-debug.log
@@ -1,18 +1,20 @@
 0 info it worked if it ends with ok
-1 verbose cli [ '/usr/local/bin/node', '/usr/local/bin/npm', 'start' ]
+1 verbose cli [ '/usr/local/bin/node', '/usr/local/bin/npm', 'install' ]
 2 info using npm@2.11.2
 3 info using node@v0.12.5
-4 verbose stack Error: ENOENT, open '/home/huginn/Development/TeamKraken/package.json'
-4 verbose stack     at Error (native)
-5 verbose cwd /home/huginn/Development/TeamKraken
-6 error Linux 3.19.0-22-generic
-7 error argv "/usr/local/bin/node" "/usr/local/bin/npm" "start"
-8 error node v0.12.5
-9 error npm  v2.11.2
-10 error path /home/huginn/Development/TeamKraken/package.json
-11 error code ENOENT
-12 error errno -2
-13 error enoent ENOENT, open '/home/huginn/Development/TeamKraken/package.json'
-13 error enoent This is most likely not a problem with npm itself
-13 error enoent and is related to npm not being able to find a file.
-14 verbose exit [ -2, true ]
+4 verbose readDependencies loading dependencies from /home/huginn/Development/TeamKraken/package.json
+5 error install Couldn't read dependencies
+6 verbose stack Error: ENOENT, open '/home/huginn/Development/TeamKraken/package.json'
+6 verbose stack     at Error (native)
+7 verbose cwd /home/huginn/Development/TeamKraken
+8 error Linux 3.19.0-22-generic
+9 error argv "/usr/local/bin/node" "/usr/local/bin/npm" "install"
+10 error node v0.12.5
+11 error npm  v2.11.2
+12 error path /home/huginn/Development/TeamKraken/package.json
+13 error code ENOPACKAGEJSON
+14 error errno -2
+15 error package.json ENOENT, open '/home/huginn/Development/TeamKraken/package.json'
+15 error package.json This is most likely not a problem with npm itself.
+15 error package.json npm can't find a package.json file in your current directory.
+16 verbose exit [ -2, true ]
